requirements.txt
asgiref==3.9.1
dj-database-url==3.0.1
Django==5.2.4
django-ckeditor==6.7.3
django-js-asset==3.1.2
mysqlclient==2.2.7
pillow==11.3.0
sqlparse==0.5.3
tzdata==2025.2
